Moving building code to new land_use game part
